added tox.ini, fixed sum_of_nth_terms.py for 2.7, added dockstring to parenthetics(), changed file names
